Make Goals and Program responsive in index.html
Here is a live version of the change: https://codepen.io/A_A/pen/gxdxoO?editors=1000
(Resize the window to see the difference)

Also, before "Goals" was left-aligned and "Program" was centered. Now to make it consistent (especially for small screens it looked weird) I centered both. Maybe it would fit better to left-align both with a little indent?